% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governor.R
\name{gov_enable}
\alias{gov_enable}
\alias{gov_disable}
\title{Disable/enable a governor}
\usage{
gov_enable(gov)

gov_disable(gov)
}
\arguments{
\item{gov}{object created by \code{gov_init()}}
}
\value{
None.
}
\description{
When disabled a governor always returns immediately without any waiting
}
\examples{
gov <- gov_init(1/30)
gov_disable(gov)
gov_enable(gov)
}
