% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{timer_enable}
\alias{timer_enable}
\alias{timer_disable}
\title{Disable/enable a timer}
\usage{
timer_enable(timer)

timer_disable(timer)
}
\arguments{
\item{timer}{timer object returned by \code{timer_init()}}
}
\value{
None.
}
\description{
When disabled a timer always immediately returns \code{FALSE}
}
\examples{
timer <- timer_init(1)
timer_disable(timer)
timer_enable(timer)
}
